KW-442/chore: rename(did-service -> did-agent)
