+++ .github/workflows/develop_ci_test.yml
@@ -9,7 +9,7 @@ permissions:
   packages: write
 
 env:
-  DOCKER_IMAGE_NAME: sunwoo11/did-service
+  DOCKER_IMAGE_NAME: sunwoo11/did-agent
   DOCKER_TAG: latest
 
 jobs:
