+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'did-service'
+rootProject.name = 'did-agent'
