+++ src/main/java/com/doubleo/didagent/controller/DidController.java
@@ -1,6 +1,6 @@
-package com.doubleo.didservice.controller;
+package com.doubleo.didagent.controller;
 
-import com.doubleo.didservice.dto.response.DidCreateResponse;
+import com.doubleo.didagent.dto.response.DidCreateResponse;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
