+++ src/main/java/com/doubleo/didagent/dto/response/DidCreateResponse.java
@@ -1,0 +1,3 @@
+package com.doubleo.didagent.dto.response;
+
+public record DidCreateResponse() {}
