+++ src/main/java/com/doubleo/didagent/global/exception/GlobalExceptionHandler.java
@@ -1,8 +1,8 @@
-package com.doubleo.didservice.global.exception;
+package com.doubleo.didagent.global.exception;
 
-import com.doubleo.memberservice.global.exception.errorcode.BaseErrorCode;
-import com.doubleo.memberservice.global.exception.errorcode.GlobalErrorCode;
-import com.doubleo.memberservice.global.response.CommonResponse;
+import com.doubleo.didagent.global.exception.errorcode.BaseErrorCode;
+import com.doubleo.didagent.global.exception.errorcode.GlobalErrorCode;
+import com.doubleo.didagent.global.response.CommonResponse;
 import lombok.SneakyThrows;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
