+++ src/main/java/com/doubleo/didagent/global/exception/errorcode/DIDErrorCode.java
@@ -1,4 +1,4 @@
-package com.doubleo.didservice.global.exception.errorcode;
+package com.doubleo.didagent.global.exception.errorcode;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import org.springframework.http.HttpStatus;
 @Getter
 @AllArgsConstructor
 public enum DIDErrorCode implements BaseErrorCode {
-    SAMPLE_ERROR(HttpStatus.NOT_FOUND, "DID Service API Sample Error"),
+    SAMPLE_ERROR(HttpStatus.NOT_FOUND, "DID Agent API Sample Error"),
     ;
 
     private final HttpStatus httpStatus;
