+++ src/main/java/com/doubleo/didagent/global/response/CommonResponseAdvice.java
@@ -1,4 +1,4 @@
-package com.doubleo.didservice.global.response;
+package com.doubleo.didagent.global.response;
 
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.core.MethodParameter;
