--- src/main/java/com/doubleo/didservice/dto/response/DidCreateResponse.java
@@ -1,3 +1,0 @@
-package com.doubleo.didservice.dto.response;
-
-public record DidCreateResponse() {}
